Made huge changes to how everything works
Since minimax was being severely bottlenecked because of the way the app works, I changed how the pieces are represented, and how the board is seen by the game. This allows me to have much quicker lookups for getting available moves which before was slowing everything down. I have worked many hours on this today and need to rest that's why not every feature is implemeneted... Maybe later today or tomorrow