Sped up AI
When looking for available moves from the AI, instead of looking for check after every move, I changed it to allow illegal moves into the available moves array, however if the king ends up getting taken it will completely skip the move, making it impossible for the AI to make an illegal move. This is much faster than searching if you've put yourself in check. For the human player, everything works as before as the performance is totally fine. Also changed the way the previous castle properties were stored which saves a little bit of times as well.